+++ tests/bootstrap.php
@@ -11,3 +11,5 @@ wget http://getcomposer.org/composer.phar
 php composer.phar install
 ");
 }
+
+