+++ tests/getCoreTest.php
@@ -1,5 +1,7 @@
 <?php
 
+# To test with a different ROOT, uncomment the below
+# define("ROOT", "/home/ubuntu/Development/zboota-server"); // Development ROOT
 require_once '/etc/zboota-server-config.php';
 require_once ROOT.'/lib/getCore.php';